+++ Controls-demo/operations/SelectedCountConfig/Index.ts
@@ -1,8 +1,8 @@
-import {Control, TemplateFunction, IControlOptions} from 'UI/Base';
+import {Control} from 'UI/Base';
 import * as template from 'wml!Controls-demo/operations/SelectedCountConfig/SelectedCountConfig';
 import {Memory} from 'Types/source';
 import {default as CountSource} from 'Controls-demo/operations/resources/Source';
-import {Gadgets} from 'Controls-demo/treeGridNew/DemoHelpers/DataCatalog';
+import { Flat } from 'Controls-demo/treeGridNew/DemoHelpers/Data/Flat';
 
 export default class extends Control {
     _template = template;
@@ -17,14 +17,14 @@ export default class extends Control {
         this._selectedKeys = [];
         this._excludedKeys = [];
         this._filter = {};
-        this._gridColumns = Gadgets.getGridColumnsForFlat();
+        this._gridColumns = Flat.getColumns();
         this._viewSource = new Memory({
-            keyProperty: 'id',
-            data: Gadgets.getFlatData()
+            keyProperty: 'key',
+            data: Flat.getData()
         });
         this._selectedCountConfig = {
             rpc: new CountSource({
-                data: Gadgets.getFlatData()
+                data: Flat.getData()
             }),
             command: 'demoGetCount',
             data: {
